<refactory> js/안에 module 파일 추가, 파일들 이름 변경, Index-Page.js 생성
[변경사항]
            -> js/Index-Page.js
js/Login.js -> js/Login-Page.js
js/CallBack.js -> js/NaverCallBack-Page.js
js/Logout.js -> js/moudle/Logout.js
js/constants.js -> js/moudle/constants.js
js/url_list.js -> js/moudle/url_list.js

refactory를 하면서 .js 파일을 추가로 더 생성을 하게 되었고 그 결과 많은 파일들로 인해서 파일을 찾는 효율이 줄어 들었고 파일 이름이
원래 부터 직관적이지 못해서 rename하게 되었습니다
