+++ apps/web-login/CallBack.html
@@ -16,7 +16,7 @@
   <body>
     <div>
       <button id="naver-logout-btn" , class="btn">Log out</button>
-      <script type="module" src="js/CallBack.js"></script>
+      <script type="module" src="js/NaverCallBack-Page.js"></script>
     </div>
   </body>
 </html>
