+++ apps/web-login/login.html
@@ -58,7 +58,7 @@
           crossorigin="anonymous"
         ></script>
         <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
-        <script type="module" src="js/Login.js"></script>
+        <script type="module" src="js/Login-Page.js"></script>
         <script type="script" src="./js/constants.js"></script>
         <!-- <script src="js/constants.js"></script> -->
       </div>
